monitoring-agent/requirements.txt
# DriftShield Monitoring Agent - Python Dependencies

# Core
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Statistics & Drift Detection
scipy>=1.11.0

# Solana
solana>=0.30.0
solders>=0.18.0
anchorpy>=0.18.0

# Storage
boto3>=1.28.0  # For S3 integration
cryptography>=41.0.0

# Monitoring & Logging
loguru>=0.7.0

# CLI
click>=8.1.0
rich>=13.0.0  # For beautiful terminal output
tabulate>=0.9.0

# HTTP
httpx>=0.24.0  # Modern async HTTP client
aiohttp>=3.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Optional ML Frameworks (install as needed)
# torch>=2.0.0  # PyTorch
# tensorflow>=2.13.0  # TensorFlow
# xgboost>=1.7.0  # XGBoost
# lightgbm>=4.0.0  # LightGBM
# transformers>=4.30.0  # Hugging Face

